break Prepend classes into files; update tests; fix docs
